move files from script/ to test/pipeline/
