+++ test/pipeline/alternativeAction/types.ts
@@ -3,8 +3,8 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 import { Raw } from '@vscode/prompt-tsx';
-import { IAlternativeAction, NextEditTelemetryStatus } from '../../src/extension/inlineEdits/node/nextEditProviderTelemetry';
-import { LogEntry } from '../../src/platform/workspaceRecorder/common/workspaceLog';
+import { IAlternativeAction, NextEditTelemetryStatus } from '../../../src/extension/inlineEdits/node/nextEditProviderTelemetry';
+import { LogEntry } from '../../../src/platform/workspaceRecorder/common/workspaceLog';
 import { ISerializedEdit } from '../logRecordingTypes';
 
 export type IStringReplacement = [start: number, endEx: number, text: string];
@@ -50,8 +50,8 @@ export namespace SuggestedEdit {
 
 export namespace Scoring {
 	export type t = {
-		"$web-editor.format-json": true;
-		"$web-editor.default-url": "https://microsoft.github.io/vscode-workbench-recorder-viewer/?editRating";
+		'$web-editor.format-json': true;
+		'$web-editor.default-url': 'https://microsoft.github.io/vscode-workbench-recorder-viewer/?editRating';
 		edits: SuggestedEdit.t[];
 		scoringContext: {
 			kind: 'recording';
@@ -61,8 +61,8 @@ export namespace Scoring {
 
 	export function create(recording: Recording.t, edits: SuggestedEdit.t[]): Scoring.t {
 		return {
-			"$web-editor.format-json": true,
-			"$web-editor.default-url": "https://microsoft.github.io/vscode-workbench-recorder-viewer/?editRating",
+			'$web-editor.format-json': true,
+			'$web-editor.default-url': 'https://microsoft.github.io/vscode-workbench-recorder-viewer/?editRating',
 			edits,
 			scoringContext: {
 				kind: 'recording',
