+++ test/pipeline/alternativeAction/util.ts
